get disk sizing right
2001-11-09  Kevin Vandersloot <kfv101@psu.edu>

        * callbacks.c: get disk sizing right

        * util.c (get_size_string): change to use gfloat since
        there were issues passing in gints

+++ src/util.h
@@ -7,5 +7,5 @@
 /* -1 means wrong password, 0 means another error, 1 means great */
 gint 		su_run_with_password (gchar *exec_path, gchar *password);
 
-gchar*		get_size_string (gint size);
+gchar*		get_size_string (gfloat size);
 #endif
